// app/(apps)/disperse/types.ts
export type TDisperseInput = {
	receiver: TInputAddressLike;
	value: TAmountInputElement;
	UUID: string;
};